django_project/pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = "Django base project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2-binary = "^2.9"
requests = "^2.31"
djangorestframework = "^3.15"
django-filter = "^24.2"
drf-spectacular = "^0.27"
aiohttp = "^3.9"
pytest = "^8.0"
pytest-django = "^4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
